requirements.txt
Django==3.2.3
django-model-utils==4.1.1
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.7.1
phonenumbers